auth/backend: move submodules into a modules folder
