+++ src/backend/src/auth/auth.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { LocalAuthModule } from './local/local-auth.module';
+import { LocalAuthModule } from './modules/local/local-auth.module';
 import { RouterModule } from '@nestjs/core';
 import { routes } from './routes';
-import { AuthUtilsModule } from './utils/auth-utils.module';
-import { JwtAuthModule } from './jwt/jwt-auth.module';
-import { OauthModule } from './oauth/oauth.module';
+import { AuthUtilsModule } from './modules/utils/auth-utils.module';
+import { JwtAuthModule } from './modules/jwt/jwt-auth.module';
+import { OauthModule } from './modules/oauth/oauth.module';
 
 @Module({
   imports: [
