+++ src/backend/src/auth/index.ts
@@ -1,2 +1,2 @@
-export * from './jwt';
-export * from './decorator';
+export * from './modules/jwt';
+export * from './decorators';
