+++ src/backend/src/auth/modules/jwt/strategy/index.ts
