+++ src/backend/src/auth/modules/jwt/strategy/jwt.strategy.ts
