+++ src/backend/src/auth/modules/utils/auth-utils.service.spec.ts
